edit done

+++ src/app/wallet/wallet-delete/wallet-delete.component.ts
@@ -4,6 +4,7 @@ import {ActivatedRoute, ParamMap, Router} from '@angular/router';
 import {FnParam} from '@angular/compiler/src/output/output_ast';
 import {WalletService} from '../../service/wallet/wallet.service';
 import Swal from 'sweetalert2';
+import {AuthencicationService} from '../../service/auth/authencication.service';
 
 @Component({
   selector: 'app-wallet-delete',
@@ -11,10 +12,13 @@ import Swal from 'sweetalert2';
   styleUrls: ['./wallet-delete.component.css']
 })
 export class WalletDeleteComponent implements OnInit {
+  idUser: number;
 
   constructor(private walletService: WalletService,
               private activatedRoute: ActivatedRoute,
-              private router: Router  ) {
+              private authService: AuthencicationService,
+              private router: Router) {
+    this.idUser = authService.currentUserValue.id;
     this.activatedRoute.paramMap.subscribe((paramMap: ParamMap) => {
       const id = paramMap.get('id');
       this.deleteWallet(id);
@@ -23,9 +27,27 @@ export class WalletDeleteComponent implements OnInit {
 
   ngOnInit() {
   }
+
   deleteWallet(id) {
-    this.walletService.delete(id).subscribe(() => {
-      this.router.navigateByUrl('/wallet/list');
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You won\'t be able to revert this!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.walletService.delete(id).subscribe(() => {
+          this.router.navigate(['/wallet/list', this.idUser]);
+        });
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        );
+      }
     });
   }
 }
